Move chiquito circuits to their own files
