+++ zkevm-circuits/src/bytecode_circuit.rs
@@ -1,5 +1,6 @@
 //! The bytecode circuit implementation.
 
+mod bytecode_chiquito;
 /// Bytecode unroller
 pub mod bytecode_unroller;
 /// Bytecode circuit
@@ -14,4 +15,5 @@ pub use dev::BytecodeCircuit as TestBytecodeCircuit;
 #[cfg(any(feature = "test", test))]
 pub mod dev;
 pub(crate) mod param;
+mod push_data_chiquito;
 mod wit_gen;
